Ability to skip individual git sub-commands
Driving use-case:

    git config --local pdp.skip-sync-cmds fsck

Takes arbitrary white-space separated list of sub-commands; does assume
first parameter after `--` is the command-line, not more git top-level
options.

For sub-commands, skip a repo by explicit testing if it's a repo, not by
failing to find remotes.  We do want to be able to fsck a remote-less
repo.