Add -A/--continue-after, support bare dirnames
The -C/--continue option took a fully-qualified path to a directory;
that's still the ideal, but also now support bare dirnames.

Add -A/--continue-after to go as far forward as one repo and then skip
that one _too_ before continuing to process; this lets us acknowledge
one temporarily broken repo without having to mark it as non-syncing and
without having to find it in a list and use -C on the next repo.